WIP: serializer for converting Halo API response to Zendesk API response.
This will need additional work to map the `id` field to the appropriate Zendesk ID. ALso, it assumes that current DBT usage of the Zendesk API only uses the `email`, `name`, and `id` fields of the Zendesk API User object. If it transpires that other fields with a Halo equivalent are used, the appropriate field and mapping should be added.