Remove default `dict` values in method arguments.
Assigning a mutable thing like a `dict` or `list` as a default value for a method argument will result in the same dict/list being used each time the application runs. In the case of a Django app, that means from the time when the server (e.g. gunicorn) loads the application, as the Django app is never normally unloaded between requests.

This can have extremely undesirable consequences and should always be avoided, even in a case like this where we never expect the method to be called without a value for that argument.