TS2-139: Support tags being set via the Zendesk api (#15)
* Remove default `dict` values in method arguments.

Assigning a mutable thing like a `dict` or `list` as a default value for a method argument will result in the same dict/list being used each time the application runs. In the case of a Django app, that means from the time when the server (e.g. gunicorn) loads the application, as the Django app is never normally unloaded between requests.

This can have extremely undesirable consequences and should always be avoided, even in a case like this where we never expect the method to be called without a value for that argument.

* Add a Zendesk ticket creation request fixture, extracted from an actual request sent by Data Workspace running locally.

* Test for Zendesk tags data being correctly transformed to the `dict` structure required for Halo.

* Transform Zendesk tags to the `dict` structure required for Halo.

* Add a Halo response, but with most fields from an actual response removed. This is the bare minimum needed for now, rather than a complete response, as I just want to get this tag stuff passing :-)

* Test for Halo tags data being correctly transformed to the `list` structure required for Zendesk.

* Transform Halo tags to the `list` structure required for Zendesk.

* Turns out Zendesk's `tags` field is just a `list` of `str` so we don't need the dataclass for them.